refactor(bundle-helpers): move shared YAML output helpers to cmd/output.go
encodeAndNormalizeYAML and normalizeYAMLOutput were in fix_descriptors.go
but also called by patch_csv.go; move them to a dedicated file to signal
that they are shared output utilities for the cmd package.

Prompt: commit this (AI-assisted)
